refactor dpm pm quality fairness routes
