+++ src/app/router_registry.py
@@ -15,6 +15,9 @@ from app.routers.dpm_command_center_outcome_reviews import (
 from app.routers.dpm_command_center_pm_quality import (
     router as dpm_command_center_pm_quality_router,
 )
+from app.routers.dpm_command_center_pm_quality_fairness import (
+    router as dpm_command_center_pm_quality_fairness_router,
+)
 from app.routers.dpm_construction import router as dpm_construction_router
 from app.routers.dpm_proof_packs import router as dpm_proof_packs_router
 from app.routers.dpm_wave_campaign_definitions import (
@@ -75,6 +78,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_router)
     app.include_router(dpm_command_center_outcome_reviews_router)
     app.include_router(dpm_command_center_pm_quality_router)
+    app.include_router(dpm_command_center_pm_quality_fairness_router)
     app.include_router(dpm_wave_campaign_definitions_router)
     app.include_router(dpm_wave_campaign_workflow_router)
     app.include_router(dpm_construction_router)
