+++ src/app/routers/dpm_command_center_pm_quality.py
@@ -76,114 +76,6 @@ async def create_pm_operating_quality_score_run(
 
 
 @router.post(
-    "/pm-operating-quality/fairness-analyses/preview",
-    response_model=DpmPmOperatingQualityGatewayResponse,
-    summary="Preview PM operating quality fairness analysis",
-    description=(
-        "What: previews a manage-owned PM operating quality fairness analysis over persisted "
-        "score runs and source-defined segment references. When: use this from governance and "
-        "evidence review views to inspect Manage-published segment posture before any broader "
-        "PM-quality operating review. How: Gateway forwards the payload unchanged and preserves "
-        "Manage state, segment results, source refs, reason codes, blocked actions, and forbidden "
-        "uses without discovering segments, calculating segment averages or score spread, "
-        "inferring protected classes, ranking PMs, administering HR/compensation/conduct actions, "
-        "approving trades, contacting clients, routing orders, or claiming execution."
-    ),
-)
-async def preview_pm_operating_quality_fairness_analysis(
-    request: DpmPmOperatingQualityForwardRequest,
-) -> DpmPmOperatingQualityGatewayResponse:
-    return await dpm_command_center_service().preview_pm_operating_quality_fairness_analysis(
-        body=request.body,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/pm-operating-quality/fairness-analyses",
-    response_model=DpmPmOperatingQualityGatewayResponse,
-    summary="Create PM operating quality fairness analysis",
-    description=(
-        "What: creates a persisted manage-owned PM operating quality fairness analysis over "
-        "persisted score runs and source-defined segment references. When: use this only after "
-        "the bank has approved the evidence posture for PM-quality governance review. How: "
-        "Gateway forwards the payload unchanged and preserves Manage state, segment results, "
-        "source refs, reason codes, blocked actions, and forbidden uses without discovering "
-        "segments, calculating fairness spread, inferring protected classes, ranking PMs, "
-        "administering HR/compensation/conduct actions, approving trades, contacting clients, "
-        "routing orders, or claiming execution."
-    ),
-)
-async def create_pm_operating_quality_fairness_analysis(
-    request: DpmPmOperatingQualityForwardRequest,
-) -> DpmPmOperatingQualityGatewayResponse:
-    return await dpm_command_center_service().create_pm_operating_quality_fairness_analysis(
-        body=request.body,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/pm-operating-quality/fairness-analyses",
-    response_model=DpmPmOperatingQualityGatewayResponse,
-    summary="List PM operating quality fairness analyses",
-    description=(
-        "What: lists persisted manage-owned PM operating quality fairness analyses. When: use "
-        "for Workbench governance evidence queues and audit posture. How: Gateway forwards "
-        "filters to Manage and preserves stored analysis payloads without calculating fairness "
-        "spread, discovering segments, inferring protected classes, ranking PMs, or converting "
-        "analysis posture into HR, compensation, conduct, client-contact, approval, execution, "
-        "or OMS decisions."
-    ),
-)
-async def list_pm_operating_quality_fairness_analyses(
-    policy_id: str | None = Query(default=None, description="Optional policy id filter."),
-    policy_version: str | None = Query(default=None, description="Optional policy version filter."),
-    as_of_date: str | None = Query(default=None, description="Optional business as-of date."),
-    state: str | None = Query(default=None, description="Optional manage-published state filter."),
-    limit: int = Query(default=50, ge=1, le=100, description="Maximum analyses to return."),
-    offset: int = Query(default=0, ge=0, description="Rows to skip."),
-) -> DpmPmOperatingQualityGatewayResponse:
-    return await dpm_command_center_service().list_pm_operating_quality_fairness_analyses(
-        filters={
-            "policy_id": policy_id,
-            "policy_version": policy_version,
-            "as_of_date": as_of_date,
-            "state": state,
-            "limit": limit,
-            "offset": offset,
-        },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/pm-operating-quality/fairness-analyses/{fairness_analysis_id}",
-    response_model=DpmPmOperatingQualityGatewayResponse,
-    summary="Get PM operating quality fairness analysis",
-    description=(
-        "What: returns one persisted manage-owned PM operating quality fairness analysis. When: "
-        "use for audit drill-down, governance evidence inspection, and Workbench read-only "
-        "fairness posture. How: Gateway retrieves Manage truth and preserves segment results, "
-        "source refs, reason codes, blocked actions, and forbidden uses without recalculating "
-        "fairness spread, inferring protected classes, ranking PMs, or creating HR, "
-        "compensation, conduct, client-contact, approval, execution, or OMS decisions."
-    ),
-)
-async def get_pm_operating_quality_fairness_analysis(
-    fairness_analysis_id: str = Path(
-        ...,
-        description="Manage-owned PM operating quality fairness-analysis identifier.",
-        examples=["pmq_fair_001"],
-    ),
-) -> DpmPmOperatingQualityGatewayResponse:
-    return await dpm_command_center_service().get_pm_operating_quality_fairness_analysis(
-        fairness_analysis_id=fairness_analysis_id,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
     "/pm-operating-quality/review-actions/preview",
     response_model=DpmPmOperatingQualityGatewayResponse,
     summary="Preview PM operating quality review action",
